requirements.txt
# LangChain Core
langchain
langchain-core
langchain_community

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# ollama
ollama
streamlit
langchain-ollama


